[Fix] jdk upgrade
