+++ app/src/main/java/com/example/netflix_clone/app/NetflixApp.kt
@@ -1,6 +1,7 @@
 package com.example.netflix_clone.app
 
 import com.example.base.app.BaseApp
+import com.example.netflix_clone.di.DaggerAppComponent
 import dagger.android.AndroidInjector
 import dagger.android.DaggerApplication
