+++ build.gradle.kts
@@ -2,6 +2,7 @@
 plugins {
     id("com.android.application") version "8.2.0" apply false
     id("org.jetbrains.kotlin.android") version "1.8.10" apply false
+    id ("androidx.navigation.safeargs") version "2.5.0" apply false
 }
 buildscript{
     repositories {
@@ -9,7 +10,6 @@ buildscript{
         mavenCentral()
     }
     dependencies {
-        classpath ("androidx.navigation:navigation-safe-args-gradle-plugin:2.7.6")
         classpath ("com.android.tools.build:gradle:8.2.0")
         // Hilt
 //        classpath ("com.google.dagger:hilt-android-gradle-plugin:2.44")
